+++ CMakeLists.txt
@@ -322,7 +322,16 @@ else()
   add_subdirectory(3party/freetype)
   add_subdirectory(3party/icu)
 endif()
-add_subdirectory(3party/jansson)
+
+set(JANSSON_EXAMPLES OFF)
+set(JANSSON_BUILD_DOCS OFF)
+set(JANSSON_WITHOUT_TESTS ON)
+set(JANSSON_INSTALL OFF)
+find_package_or_fallback_to_3party(Jansson 2.14.0 3party/jansson/jansson/)
+if (NOT Jansson_FOUND)
+  include_directories("${PROJECT_BINARY_DIR}/3party/jansson/jansson/include")
+endif()
+
 add_subdirectory(3party/liboauthcpp)
 add_subdirectory(3party/minizip)
 add_subdirectory(3party/opening_hours)
@@ -356,6 +365,7 @@ add_subdirectory(geometry)
 add_subdirectory(indexer)
 add_subdirectory(kml)
 add_subdirectory(map)
+add_subdirectory(cppjansson)
 add_subdirectory(platform)
 add_subdirectory(routing)
 add_subdirectory(routing_common)
