+++ cmake/FindJansson.cmake
@@ -1,0 +1,14 @@
+if (Jansson_INCLUDE_DIRS AND Jansson_LIBRARY AND Jansson_VERSION)
+  set(Jansson_FOUND TRUE)
+else ()
+  find_path(Jansson_INCLUDE_DIRS NAMES jansson.h)
+  find_library(Jansson_LIBRARY NAMES jansson)
+
+  if (Jansson_INCLUDE_DIRS AND Jansson_LIBRARY)
+    set(regex_jansson_version "#define[ \t]+JANSSON_VERSION[ \t]+[\"]([^\"]+)[\"]")
+    file(STRINGS "${Jansson_INCLUDE_DIRS}/jansson.h" Jansson_VERSION REGEX "${regex_jansson_version}")
+    string(REGEX REPLACE "${regex_jansson_version}" "\\1" Jansson_VERSION "${Jansson_VERSION}")
+    unset(regex_jansson_version)
+    set(Jansson_FOUND TRUE)
+  endif ()
+endif()
