+++ drape_frontend/color_constants.cpp
@@ -11,7 +11,7 @@
 #include "base/assert.hpp"
 #include "base/string_utils.hpp"
 
-#include "3party/jansson/myjansson.hpp"
+#include "cppjansson/cppjansson.hpp"
 
 #include <fstream>
