+++ generator/hierarchy_entry.cpp
@@ -14,7 +14,7 @@
 #include <tuple>
 #include <unordered_map>
 
-#include "3party/jansson/myjansson.hpp"
+#include "cppjansson/cppjansson.hpp"
 
 namespace
 {
