+++ platform/platform_tests/CMakeLists.txt
@@ -19,4 +19,5 @@ omim_add_test_with_qt_event_loop(${PROJECT_NAME} ${SRC})
 target_link_libraries(${PROJECT_NAME}
   platform_tests_support
   platform
+  cppjansson
 )
