+++ platform/servers_list.cpp
@@ -6,7 +6,7 @@
 #include "base/logging.hpp"
 #include "base/assert.hpp"
 
-#include "3party/jansson/myjansson.hpp"
+#include "cppjansson/cppjansson.hpp"
 
 namespace downloader
 {
