+++ storage/diff_scheme/diff_scheme_loader.cpp
@@ -11,7 +11,7 @@
 #include <unordered_map>
 #include <utility>
 
-#include "3party/jansson/myjansson.hpp"
+#include "cppjansson/cppjansson.hpp"
 
 #include "private.h"
