+++ storage/storage.cpp
@@ -27,7 +27,7 @@
 
 #include "defines.hpp"
 
-#include "3party/jansson/myjansson.hpp"
+#include "cppjansson/cppjansson.hpp"
 
 #include <algorithm>
 #include <sstream>
