+++ xcode/omim.xcworkspace/contents.xcworkspacedata
@@ -10,6 +10,9 @@
    <FileRef
       location = "container:common.xcconfig">
    </FileRef>
+   <FileRef
+      location = "container:cppjansson/cppjansson.xcodeproj">
+   </FileRef>
    <Group
       location = "group:../std"
       name = "std">
